Updated project creation.
Changed the way in which the project creation screen is shown, it used to use a modal that popped up and blurred the background, this worked on chrome and firefox, but failed on safari and IE, so removed the modal but kept the hiding/showing div element of it. Looks miles better on mobile, could do with work on desktop but it is fine.